pyproject.toml
[project]
name = "vision-inspection-system"
version = "1.0.0"
description = "AI-powered visual inspection system with dual-agent architecture for defect detection and safety analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Vision Inspection Team", email = "team@example.com" }
]
keywords = [
    "vision",
    "inspection",
    "ai",
    "defect-detection",
    "safety",
    "langchain",
    "langgraph",
    "vlm",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    # Core
    "requests>=2.32.5",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # LangChain & LangGraph
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.28",
    "langsmith>=0.1.0",
    # Database
    "sqlalchemy>=2.0.0",
    # UI
    "streamlit>=1.29.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    # PDF Generation
    "reportlab>=4.0.0",
    # Image Processing
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    # Logging
    "colorlog>=6.7.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
"Homepage" = "https://github.com/example/vision-inspection-system"
"Bug Tracker" = "https://github.com/example/vision-inspection-system/issues"
"Documentation" = "https://github.com/example/vision-inspection-system#readme"

[project.scripts]
vision-inspect = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "app", "utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py311"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
